Make all tasks depend on toolchain version (untested)
